+++ aqua/checks.py
@@ -17,11 +17,13 @@
 import logging
 
 from functools import wraps
-from typing import Callable
+from typing import Callable, Literal
 
 from telegram import Update, Chat
+from telegram.ext.callbackcontext import CallbackContext
 
 from aqua.constants import USER_CHAT_ID
+from aqua.utils import logged_send_message
 
 
 def is_authorized(update: Update) -> bool:
@@ -103,3 +105,53 @@ def authorize(command: Callable) -> Callable:
         return
 
     return wrapper
+
+
+def ensure_context_number_args(number_args: int, comparison_method: Literal['min', 'exact', 'max']) -> Callable:
+    """
+    Ensures the correct number of arguments were passed to the command's context.
+
+    Parameters
+    ----------
+    number_args : int
+        The number of arguments.
+    comparison_method: Literal['min', 'exact', 'max']
+        The method which will be used to check if the number of arguments are enough. For
+        example, if `comparison_method` is 'min' and `number_args` is 5, then the user must pass AT
+        LEAST 5 arguments, otherwise function will not be executed.
+
+    Returns
+    -------
+    Callable
+        A function that executes the original function if `number_args` is enough according
+        to `comparison_method`, else a function that does nothing.
+    """
+    message_join = {'min': 'at least ', 'exact': '', 'max': 'at most '}[comparison_method]
+
+    def decorator(function: Callable):
+        @wraps(function)
+        def wrapper(*args, **kwargs):
+            for arg in args:
+                if type(arg) == Update:
+                    update = arg
+                if type(arg) == CallbackContext:
+                    context = arg
+                    len_args = len(arg.args)
+                    if (
+                        (comparison_method == 'min' and len_args >= number_args)
+                        or (comparison_method == 'exact' and len_args == number_args)
+                        or (comparison_method == 'max' and len_args <= number_args)
+                    ):
+                        return function(*args, **kwargs)
+
+            logging.warn(
+                f'Function {function.__name__} was called with {len_args} arguments, but expected {number_args}. '
+                f'Comparison method used was {comparison_method} - Not executing function.'
+            )
+            logged_send_message(
+                update, context,
+                f'Sorry, this command expects {message_join}{number_args} arguments, '
+                f'but received {len(context.args)} instead.'
+            )
+        return wrapper
+    return decorator