+++ aqua/extensions/commands/_invalid.py
@@ -17,11 +17,12 @@
 from telegram import Update
 from telegram.ext.callbackcontext import CallbackContext
 
-from aqua.checks import authorize
+from aqua.checks import authorize, ensure_context_number_args
 from aqua.utils import logged_send_message
 
 
 @authorize
+@ensure_context_number_args(0, 'min')
 def _invalid(update: Update, context: CallbackContext) -> None:
     msg = 'Sorry, it seems you tried to use a command that is either invalid or not correctly abbreviated.\n\n'
     msg += 'Check a list of available commands with: /help'