+++ tests/test_static_analysis.py
@@ -36,20 +36,15 @@ def test_auths_for_all_aqua_extensions():
             src_code = f.read()
 
         parsed_ast = ast.parse(src_code)
-
-        found_auth = False
+        results[command_to_seek] = 'FAILED'
         for node in ast.walk(parsed_ast):
             if isinstance(node, ast.FunctionDef) and node.name == command_to_seek:
                 # If the command was found, then it MUST have auth
                 for child_node in ast.iter_child_nodes(node):
                     if isinstance(child_node, ast.Name) and child_node.id == 'authorize':
-                        found_auth = True
                         results[command_to_seek] = 'OK'
 
-        if not found_auth:
-            results[command_to_seek] = 'FAILED'
-            print(f'{command_to_seek} has no auth.')
-
+    print(results)
     assert 'FAILED' not in results.values()
 
 
@@ -63,7 +58,7 @@ def test_documentation_for_all_aqua_extensions():
     for _file in AQUA_EXTENSIONS_FILES:
         command_to_seek = '/' + basename(_file).replace('.py', '')
         if command_to_seek.startswith('/_') or command_to_seek == '/talk':
-            # Internal commands do not need docs, neither does the talk handler
+            # Internal commands do not need docs, neither does the talk handler.
             continue
 
         remaining_commands_to_find_in_help.add(command_to_seek)